Adjust QS battery tile padding from 2dp to 0dp
We're currently syncing the battery from the statusbar to the QS tile. While
that works, the size is just not the same as it was on stock. This adjustment
brings it closer to stock until we can come up with a more permanent fix.

See image - https://i.imgur.com/47pytg0.png

Change-Id: I93b99c84c9124a6ad157e8285263e55cf24d7c1a